pyproject.toml
[tool.poetry]
name = "hybrid-vocal-classifier"
version = "0.3.0"
description = "a Python machine learning library for animal vocalizations and bioacoustics"
authors = ["David Nicholson <nickledave@users.noreply.github.com>"]
license = "BSD"
classifiers = [
    'License :: OSI Approved :: BSD License',
    'Development Status :: 5 - Production/Stable',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: Implementation :: CPython',
]
readme = 'README.md'
homepage = 'https://github.com/NickleDave/hybrid-vocal-classifier'
packages = [
    {include = "hvc", from="src"}
]
include = [
    'src/hvc/**/*yml'
]

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
evfuncs = ">=0.3.2"
numpy = ">=1.19.2"
scipy = ">=1.6.2"
matplotlib = ">=3.4.1"
scikit-learn = ">=0.24.1"
Keras = ">=2.4.3"
tensorflow = ">=2.4.1"
PyYAML = ">=5.4.1"
joblib = ">=1.0.1"


[tool.poetry.dev-dependencies]
pytest = ">=6.2.3"
Sphinx = ">=3.5.3"
pytest-cov = "^2.11.1"
coverage = {extras = ["toml"], version = "^5.5"}
ipython = ">=7.22.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
